All public polls routes functional

+++ src/routes/polls.js
@@ -1,3 +1,5 @@
+const { validateToken, verifyTokenIsAdmin } = require('../util/auth')
+const log = require('../util/tools').getLogger()
 const Mongo = require('../db')
 
 const main = server => {
@@ -5,14 +7,20 @@ const main = server => {
   // Returns all polls currently created for the given team.
   server.get({ path: '/polls/:clientId/:teamId', version: '1' }, async (req, res, next) => {
     try {
-      const pService = Mongo.getPollService(req)
-      const polls = await pService.getPolls()
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, polls)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const pService = Mongo.getPollService(req)
+        const polls = await pService.getPolls()
+
+        res.send(200, polls)
+      }
+
       return next()
     } catch (err) {
-      console.error(err)
-      res.send(err)
+      log.error(err)
+      res.send(500, err)
       return next(err)
     }
   })
@@ -21,14 +29,21 @@ const main = server => {
   // Creates a new poll for the given team using the given NewPoll object.
   server.post({ path: '/polls/:clientId/:teamId', version: '1' }, async (req, res, next) => {
     try {
-      const pService = Mongo.getPollService(req)
-      const newPoll = await pService.createPoll(req.body)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, newPoll)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const pService = Mongo.getPollService(req)
+        const newPoll = await pService.createPoll(req.body)
+
+        if (newPoll instanceof Error) res.send(409, 'A poll already exists with the given displayName.')
+        else res.send(200, newPoll)
+      }
+
       return next()
     } catch (err) {
-      console.error(err)
-      res.send(err)
+      log.error(err)
+      res.send(500, err)
       return next(err)
     }
   })
@@ -36,21 +51,39 @@ const main = server => {
   // DELETE /api/v1/polls/:clientId/:teamId
   // Delete all polls that exist for a given Team.
   server.del({ path: '/polls/:clientId/:teamId', version: '1' }, async (req, res, next) => {
-    // Add step to verify the requesting user is an admin of sorts.
+    res.send(501, 'This endpoint has not been implemented yet.')
+    return next()
+    // try {
+    //   const validToken = await verifyTokenIsAdmin(req.headers['request-token'])
+    //   if (!validToken) res.send(401)
+    //   else {
+
+    //   }
+    // } catch (err) {
+    //   log.error(err)
+    //   res.send(500, err)
+    //   return next(err)
+    // }
   })
 
   // GET /api/v1/polls/:clientId/:teamId/:pollId
   // Get a specific poll by id.
   server.get({ path: '/polls/:clientId/:teamId/:pollId', version: '1' }, async (req, res, next) => {
     try {
-      const pService = Mongo.getPollService(req)
-      const poll = await pService.getPoll(req.params.pollId)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, poll)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const pService = Mongo.getPollService(req)
+        const poll = await pService.getPoll(req.params.pollId)
+
+        res.send(200, poll)
+      }
+
       return next()
     } catch (err) {
-      console.error(err)
-      res.send(err)
+      log.error(err)
+      res.send(500, err)
       return next(err)
     }
   })
@@ -59,14 +92,20 @@ const main = server => {
   // Update the information an existing poll.
   server.put({ path: '/polls/:clientId/:teamId/:pollId', version: '1' }, async (req, res, next) => {
     try {
-      const pService = Mongo.getPollService(req)
-      const updatedPoll = await pService.updatePoll(req.params.pollId, req.body)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, updatedPoll)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const pService = Mongo.getPollService(req)
+        const updatedPoll = await pService.updatePoll(req.params.pollId, req.body)
+
+        res.send(200, updatedPoll)
+      }
+
       return next()
     } catch (err) {
-      console.error(err)
-      res.send(err)
+      log.error(err)
+      res.send(500, err)
       return next(err)
     }
   })
@@ -75,14 +114,19 @@ const main = server => {
   // Delete an existing poll.
   server.del({ path: '/polls/:clientId/:teamId/:pollId', version: '1' }, async (req, res, next) => {
     try {
-      const pService = Mongo.getPollService(req)
-      const dRes = await pService.deletePollFromTeam(req.params.pollId)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, dRes)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const pService = Mongo.getPollService(req)
+        const dRes = await pService.deletePollFromTeam(req.params.pollId)
+
+        res.send(200, dRes)
+      }
       return next()
     } catch (err) {
-      console.error(err)
-      res.send(err)
+      log.error(err)
+      res.send(500, err)
       return next(err)
     }
   })
@@ -91,14 +135,19 @@ const main = server => {
   // Get the question for an existing Poll.
   server.get({ path: '/polls/:clientId/:teamId/:pollId/question', version: '1' }, async (req, res, next) => {
     try {
-      const pService = Mongo.getPollService(req)
-      const question = await pService.getPollQuestion(req.params.pollId)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, question)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const pService = Mongo.getPollService(req)
+        const question = await pService.getPollQuestion(req.params.pollId)
+
+        res.send(200, question)
+      }
       return next()
     } catch (err) {
-      console.error(err)
-      res.send(err)
+      log.error(err)
+      res.send(500, err)
       return next(err)
     }
   })
@@ -107,14 +156,19 @@ const main = server => {
   // Get all of the available responses for an existing poll.
   server.get({ path: '/polls/:clientId/:teamId/:pollId/responses', version: '1' }, async (req, res, next) => {
     try {
-      const pService = Mongo.getPollService(req)
-      const responses = await pService.getPollResponses(req.params.pollId)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, responses)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const pService = Mongo.getPollService(req)
+        const responses = await pService.getPollResponses(req.params.pollId)
+
+        res.send(200, responses)
+      }
       return next()
     } catch (err) {
-      console.error(err)
-      res.send(err)
+      log.error(err)
+      res.send(500, err)
       return next(err)
     }
   })
@@ -123,19 +177,24 @@ const main = server => {
   // Add a new possible response to an existing poll.
   server.post({ path: '/polls/:clientId/:teamId/:pollId/responses', version: '1' }, async (req, res, next) => {
     try {
-      const pService = Mongo.getPollService(req)
-      const status = await pService.addResponseToPoll(req.params.pollId, req.body)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      if (status.code === 409) {
-        res.send(409, status.msg)
-        return next()
-      } else {
-        res.send(200, status)
-        return next()
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const pService = Mongo.getPollService(req)
+        const status = await pService.addResponseToPoll(req.params.pollId, req.body)
+
+        if (status.code === 409) {
+          res.send(409, status.msg)
+          return next()
+        } else {
+          res.send(200, status)
+          return next()
+        }
       }
     } catch (err) {
-      console.error(err)
-      res.send(err)
+      log.error(err)
+      res.send(500, err)
       return next(err)
     }
   })
@@ -144,19 +203,24 @@ const main = server => {
   // Cast a vote on an existing poll.
   server.post({ path: '/polls/:clientId/:teamId/:pollId/vote', version: '1' }, async (req, res, next) => {
     try {
-      const pService = Mongo.getPollService(req)
-      const status = await pService.voteOnPollResponse(req.params.pollId, req.body)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      if (status.code === 404) {
-        res.send(404, status.msg)
-        return next()
-      } else {
-        res.send(200, status)
-        return next()
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const pService = Mongo.getPollService(req)
+        const status = await pService.voteOnPollResponse(req.params.pollId, req.body)
+
+        if (status.code === 404) {
+          res.send(404, status.msg)
+          return next()
+        } else {
+          res.send(200, status)
+          return next()
+        }
       }
     } catch (err) {
-      console.error(err)
-      res.send(err)
+      log.error(err)
+      res.send(500, err)
       return next(err)
     }
   })
